test: extract expect extension & types
