+++ @types/vitest/index.d.ts
@@ -1,0 +1,10 @@
+import vitest from 'vitest';
+
+interface CustomMatchers<R = unknown> {
+  anyMockFunction(): R;
+}
+
+declare module 'vitest' {
+  interface Assertion<T = any> extends CustomMatchers<T> {}
+  interface AsymmetricMatchersContaining extends CustomMatchers {}
+}
